GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - update-api-base-url-constant (refs/gitbutler/update-api-base-url-constant)
   - src/get-player-sumaries/get-player-sumaries.ts
   - src/get-player-sumaries/response.ts
   - src/config.ts

Your previous branch was: refs/heads/main

The sha for that commit was: 636fd37fb93315cac441e4673cac6430387e5fb6

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch